Support account membership events (#6)
* Update dnsimple module

* Introduce test for default message

The goal here is being able to write specs for the message formatters. I
have used the simplest one as it requires less data about the event.

* Handle account.user_invite event

* Handle account.user_invitation_accept event

* Handle account.user_invitation_revoke event

* Rely on actor to show who performs action

Instead of trying to infer it from other parts of the payload (which can
even be providing the wrong information).

* Use Account to generate members link

Instead of AccountInvitation. Account.ID and AccountInvitation.AccountID
are the same. The difference is that the account.user_remove event does
not include an AccountInvitation in its payload. With this change we
make the link generation valid for all membership related events.

* Handle account.user_remove event

* Minor cleaning up

* Use tabs also in multiline strings